refactor sorted albums component
